add connectors
