+++ routes/notes.js
@@ -1,43 +1,16 @@
 const express = require('express');
-const bodyParser = require('body-parser');
-const isDateValid = require('../validations/validationHandler');
 
-const jsonParser = bodyParser.json();
+const isDateValid = require('../validations/validationHandler');
+const { postNote, putNote, deleteNote, getNotes } = require('../connectors/notesConnetctors');
+
 const router = express.Router();
 
-let notes = [];
-let currentNote;
+router.get('/notes', getNotes);
 
-const findElement = (req, res, next) => {
-    currentNote = notes.find((note) => note.id === req.params.id);
-    if (!currentNote) {
-        return res.status(404).json({ error: { message: 'note not found' } });
-    }
-    next();
-};
+router.post('/notes', isDateValid, postNote);
 
-const putElement = (req, res, next) => {
-    notes = notes.map((note) => (note.id === req.params.id ? req.body : note));
-    next();
-};
+router.put('/notes/:id', isDateValid, putNote);
 
-router.get('/notes', (req, res) => {
-    res.send(notes);
-});
-
-router.post('/notes', [jsonParser, isDateValid], (req, res) => {
-    notes.push(req.body);
-    res.json(req.body);
-});
-
-router.put('/notes/:id', [jsonParser, findElement, isDateValid, putElement], (req, res) => {
-    res.json(req.body);
-});
-
-router.delete('/notes/:id', findElement, (req, res) => {
-    const deleteElement = notes.findIndex((note) => note.id === req.params.id);
-    notes.splice(deleteElement, 1);
-    res.json({ success: 'true', id: req.params.id });
-});
+router.delete('/notes/:id', deleteNote);
 
 module.exports = router;
